% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD.R
\name{plotD}
\alias{plotD}
\title{Plots heatmap for scaling coefficients obtained when generating PV-set for PCR or PLS}
\usage{
plotD(
  Xpv,
  colmap = colorRampPalette(c("blue", "white", "red"))(256),
  lim = c(-2, 4),
  xlab = "Components",
  ylab = "Segments",
  ...
)
}
\arguments{
\item{Xpv}{PV-set generated by `pcvpcr()` or `pcvpls()`.}

\item{colmap}{colormap - any with 256 colors.}

\item{lim}{limits for color map (smallest/largest expected value), centered around 1.}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{...}{any other parameters for method `image`}
}
\value{
No return value, just creates a plot.
}
\description{
Plots heatmap for scaling coefficients obtained when generating PV-set for PCR or PLS
}
