% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvpls.R
\name{simpls}
\alias{simpls}
\title{SIMPLS algorithm}
\usage{
simpls(X, Y, ncomp)
}
\arguments{
\item{X}{a matrix with x values (predictors)}

\item{Y}{a matrix with y values (responses)}

\item{ncomp}{number of components to calculate}
}
\value{
a list with computed weights, x- and y-loadings for PLS regression model.
}
\description{
SIMPLS algorithm for calibration of PLS model
}
\references{
[1]. S. de Jong. SIMPLS: An Alternative approach to partial least squares regression.
Chemometrics and Intelligent Laboratory Systems, 18, 1993 (251-263).
}
